Make `flepimop2._cli` semi-public
Converted the `flepimop2._cli` subpackage to semi-public via
`flepimop2.cli`, primarily to make `CliCommand` easily available for job
module developers.
