+++ docs/reference/cli.md
@@ -1,7 +1,7 @@
 # CLI
 
 ::: mkdocs-click
-    :module: flepimop2._cli._cli
+    :module: flepimop2.cli._cli
     :command: cli
     :prog_name: flepimop2
     :depth: 1
