+++ src/flepimop2/cli/__init__.py
@@ -13,18 +13,8 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-"""Tests for the `get_option` function in `flepimop2._cli._options`."""
+"""Public CLI API for external command implementations."""
 
-import pytest
+from flepimop2.cli._cli_command import CliCommand
 
-from flepimop2._cli._options import COMMON_OPTIONS, get_option
-
-
-@pytest.mark.parametrize("name", ["nope", "does not exist", "another fake option"])
-def test_key_error_when_option_not_found(name: str) -> None:
-    """Test that get_option raises KeyError for unknown option names."""
-    assert name not in COMMON_OPTIONS
-    with pytest.raises(
-        KeyError, match=rf"^\"Unknown option '{name}'\. Available options:"
-    ):
-        get_option(name)
+__all__ = ["CliCommand"]
