+++ tests/cli/_options/test_get_option.py
@@ -13,4 +13,18 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-__all__ = []
+"""Tests for the `get_option` function in `flepimop2.cli._options`."""
+
+import pytest
+
+from flepimop2.cli._options import COMMON_OPTIONS, get_option
+
+
+@pytest.mark.parametrize("name", ["nope", "does not exist", "another fake option"])
+def test_key_error_when_option_not_found(name: str) -> None:
+    """Test that get_option raises KeyError for unknown option names."""
+    assert name not in COMMON_OPTIONS
+    with pytest.raises(
+        KeyError, match=rf"^\"Unknown option '{name}'\. Available options:"
+    ):
+        get_option(name)
