+++ tests/cli/test_format_command.py
@@ -19,7 +19,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-from flepimop2._cli._cli import cli
+from flepimop2.cli._cli import cli
 from flepimop2.typing import ExitCode
