+++ tests/logging/test_logging_level_class.py
@@ -13,13 +13,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-"""Unit tests for the `LoggingLevel` class in `flepimop2._cli._logging` module."""
+"""Unit tests for the `LoggingLevel` class in `flepimop2.cli._logging` module."""
 
 import logging
 
 import pytest
 
-from flepimop2._cli._logging import LoggingLevel
+from flepimop2.cli._logging import LoggingLevel
 
 
 @pytest.mark.parametrize("verbosity", [-1, -100])
